fix cyclical dependency
